remove setavatar attribute

+++ app/Models/User.php
@@ -3,6 +3,7 @@
 namespace App\Models;
 
 use App\Handlers\ImageUploadHandler;
+use App\Models\Traits\ActiveUserHelper;
 use Illuminate\Auth\MustVerifyEmail as MustVerifyEmailTrait;
 use Illuminate\Contracts\Auth\MustVerifyEmail as MustVerifyEmailContract;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
@@ -15,9 +16,11 @@ use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Storage;
 use Intervention\Image\Facades\Image;
 
+
 class User extends Authenticatable implements MustVerifyEmailContract
 {
     use HasApiTokens, HasFactory, Notifiable, MustVerifyEmailTrait, HasRoles;
+    use ActiveUserHelper;
 
     /**
      * The attributes that are mass assignable.
@@ -83,14 +86,17 @@ class User extends Authenticatable implements MustVerifyEmailContract
 
     public function setAvatarAttribute($fileName)
     {
-        $path = public_path() . "/uploads/images/avatars/$fileName";
-        (Image::make($path))->resize(416, 416)->save();
-        $folder_name = "uploads/images/avatars/" . date("Ym/d", time());
-        $image = file_get_contents($path);
-        app('upyun')->write("$folder_name/$fileName", $image);
-        unlink($path);
-        $this->attributes['avatar'] = config('services.upyun.domain') . "$folder_name/$fileName";
+        if (!app()->runningInConsole()) {
+            $path = public_path() . "/uploads/images/avatars/$fileName";
+            (Image::make($path))->resize(416, 416)->save();
+            $folder_name = "uploads/images/avatars/" . date("Ym/d", time());
+            $image = file_get_contents($path);
+            app('upyun')->write("$folder_name/$fileName", $image);
+            unlink($path);
+            $this->attributes['avatar'] = config('services.upyun.domain') . "$folder_name/$fileName";
+        }
     }
+
     public function replyNotify($instance)
     {
         // 如果要通知的人是当前用户，就不必通知了！
